requirements.txt
flask
Flask-SQLAlchemy
mysqlclient
python-dotenv